+++ tests/_cases/DbReport.php
@@ -1,4 +1,5 @@
 <?php
+require_once dirname(__FILE__)."/../../autoload.php";
 
 class DbReport extends \koolreport\KoolReport
 {